fix(library): storage total showed ~8192 TB — cast SUM(size_bytes) to bigint
Postgres SUM(bigint) returns NUMERIC, which Prisma maps to Decimal.
Mixed into the stats builder's bigint accumulation it concatenated
digits instead of adding, blowing past MAX_SAFE_INTEGER, so the
clamp rendered as ~8192 TB for any library with 2+ resolution
buckets.

+++ apps/api/src/routes/library/libraryJobStatsRoutes.ts
@@ -25,8 +25,12 @@ export const libraryJobStatsRoutes = new Elysia()
           where: { type: "show" },
           _count: true,
         }),
+        // ::bigint — SUM(bigint) yields NUMERIC, which Prisma maps to Decimal;
+        // Decimal leaking into buildLibraryStatsResponse's bigint arithmetic
+        // corrupts the storage total (digit concatenation, then the
+        // MAX_SAFE_INTEGER clamp shows ~8192 TB).
         prisma.$queryRaw<{ resolution: number | null; size_bytes: bigint }[]>`
-          SELECT resolution, SUM(size_bytes) AS size_bytes
+          SELECT resolution, SUM(size_bytes)::bigint AS size_bytes
           FROM media_files
           GROUP BY resolution
         `,
